test: remove open-media test, and implement it into Profile0-non-strict
