+++ packages/connector/src/__tests__/Profile0-non-strict.spec.ts
@@ -238,4 +238,27 @@ describe('Profile 0 - non strict', () => {
 
 		expect(String(caughtError)).toMatch(/Unable to parse MOS reply.*Invalid timestamp/i)
 	})
+	test('requestMachineInfo - empty <time> - as returned by OpenMedia ', async () => {
+		// Prepare mock server response:
+		const mockReply = jest.fn((data) => {
+			const str = decode(data)
+			const messageID = getMessageId(str)
+			const replyMessage = xmlData.machineInfoOpenMedia
+			const repl = getXMLReply(messageID, replyMessage)
+			return encode(repl)
+		})
+		socketMockLower.mockAddReply(mockReply)
+		if (!xmlApiData.mosObj.ID) throw new Error('xmlApiData.mosObj.ID not set')
+
+		const returnedMachineInfo: IMOSListMachInfo = await mosDevice.requestMachineInfo()
+		expect(mockReply).toHaveBeenCalledTimes(1)
+		const msg = decode(mockReply.mock.calls[0][0])
+		expect(msg).toMatch(/<reqMachInfo\/>/)
+		checkMessageSnapshot(msg)
+
+		const replyMessage = { ...xmlApiData.machineInfoOpenMediaReply }
+
+		expect(returnedMachineInfo).toMatchObject(replyMessage)
+		expect(returnedMachineInfo.opTime).toBeUndefined()
+	})
 })
