+++ packages/model/src/mosTypes/__tests__/mosTime.spec.ts
@@ -12,16 +12,6 @@ describe('MosTime', () => {
 		expect(() => mosTypes.mosTime.validate(mosTime)).not.toThrowError()
 		// @ts-expect-error wrong input, but still:
 		expect(mosTypes.mosTime.valueOf(123456)).toBe(123456)
-
-		const nonStrictMosTypes = getMosTypes(false)
-		// @ts-expect-error wrong input, but happens when using OpenMedia MOS:
-		const newTime = nonStrictMosTypes.mosTime.create({})
-		newTime._mosTime = 1710748094984
-		expect(newTime).toStrictEqual({
-			_mosTime: 1710748094984,
-			_timezone: '',
-			_timezoneOffset: 0,
-		})
 	})
 	test('is', () => {
 		const mosTypes = getMosTypes(true)
@@ -126,7 +116,7 @@ describe('MosTime', () => {
 		// @ts-expect-error wrong input format, but still:
 		expect(mosTypes.mosTime.stringify('2009-04-11T14:22:07,000+05:00')).toBe('2009-04-11T14:22:07,000+05:00')
 	})
-	test('handles tricky (midnight, new year, summer time, leap year) corectly', () => {
+	test('handles tricky (midnight, new year, summer time, leap year) correctly', () => {
 		const mosTypes = getMosTypes(true)
 		function toTime(input: any) {
 			return mosTypes.mosTime.valueOf(mosTypes.mosTime.create(input))
@@ -171,5 +161,16 @@ describe('MosTime', () => {
 		// @todo: daylight savings fall
 
 		// @todo: leap year
+	})
+	test('Non expected inputs', () => {
+		const nonStrictMosTypes = getMosTypes(false)
+
+		// @ts-expect-error wrong input, but happens when using OpenMedia MOS:
+		const newTime = nonStrictMosTypes.mosTime.create({}) // This should not throw an error
+		expect(newTime).toStrictEqual({
+			_mosTime: expect.any(Number),
+			_timezone: '',
+			_timezoneOffset: 0,
+		})
 	})
 })
